build GoBack component
